### 2025-12-15 #9630 @editorialbot
[PRE REVIEW]: pyfive: A pure-Python HDF5 reader
**Submitting author:** @bnlawrence (<a href="http://orcid.org/0000-0001-9262-7860">Bryan Lawrence</a>)
**Repository:** https://github.com/ncas-cms/pyfive
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
