### 2025-12-15 871bc03
fix code assertions, fix block_size for indirect blocks, iterate over all possible blocks, add first fractal heap tests
