### 2025-12-16 99c2ea9
Some documentation and debug utiliity for the future.
