### 2025-12-16 363e46e
Update pyfive/misc_low_level.py
