### 2025-12-17 4323263
Documentation modifications required by #94
