### 2025-12-17 cc2efde
Drastic shortening as requested by JOSS
