+++ paper.md
@@ -68,8 +68,8 @@ With these extensions, coupled with thread safety, many of the limitations precl
 
 # Statement of need
 
-HDF5[^1][@FolEA11] is probably the most important data format in physical science, used across the piste. 
-It is particularly important in environmental science, particularly given the fact that netCDF4[^2][@Rew06] is HDF5 under the hood. 
+HDF5[^1][@FolEA11] is probably the most important data format in physical science.
+It is particularly important in environmental science given the fact that netCDF4[^2][@Rew06] is HDF5 under the hood. 
 From satellite missions, to climate models and radar systems, the default binary format has been HDF5 for decades. 
 While newer formats are starting to get mindshare, there are petabytes, if not exabytes, of existing HDF5, and there are still many good use-cases for creating new data in HDF5. 
 However, despite the history, there are few libraries for reading HDF5 file data that do not depend on the official HDF5 library maintained by the HDF Group, and in particular, apart from `pyfive`, in Python there are none that cover the needs of environmental science. 
@@ -78,197 +78,45 @@ Firstly, it is not thread-safe,  secondly, the code is large and complex, and sh
 Finally, the code complexity also meant that it is not suitable for developing bespoke code for data recovery in the case of partially corrupt data. 
 From a long-term curation perspective both of these last two constraints are a concern. 
 
+The issues of the dependency on a complex code maintained by one private company in the context of maintaining data access (over decades, and potentially centuries), can only be mitigated by ensuring that the data format is well documented, that data writers use only the documented features, and that public code exists which can be relatively easily maintained. 
+The HDF5group have provided good documentation for HDF5, but while there is a community of developers beyond the HDF5 group, recent events suggest that given most of those developers and their existing funding are US based, some spreading of risk would be desirable. 
+To that end, a pure Python code covering the core HDF5 features of interest to the target scientific community, which is relatively small and maintained by an international constituency provides some assurance that the community can maintain HDF5 access for the foreseeable future. 
+A pure Python code also makes it easier to develop scripts which can work around data and metadata damage should they occur, and has the additional advantage of being able to be deployed in resource or operating-system constrained environments (such as mobile).
+
+
 [^1]: https://www.hdfgroup.org/solutions/hdf5/
 [^2]: https://www.unidata.ucar.edu/software/netcdf
 
+# Current Status
+
+
 The original implementation of `pyfive` (by JH), which included all the low-level functionality to deal with the internals of an HDF5 file was developed with POSIX access in mind. 
-The recent upgrades were developed with the use-case of performant remote access to curated data as the primary motivation, but with additional motivations of having a lightweight HDF5 reader capable of deploying in resource or operating-system constrained environments (such as mobile), and one that could be maintained long-term as a reference reader for curation purposes. 
-The lightweight deployment consequences of a pure-Python HDF5 reader needs no further introduction, but as additional motivation we now expand on the issues around remote access and curation.
+The recent upgrades were developed with the use-cases of performant remote access to curated data as the primary motivation - including full support for lazy loading only parts of chunked datasets as they are needed.
 
-Thread safety has become a concern given the wide use of Dask[^3] in Python based analysis workflows, and this, coupled with a lack of user knowledge about how to efficiently use HDF5, has led to a community perception that HDF5 is not fit for remote access (especially on cloud storage). 
-Issues with thread safety arise from the underlying HDF5 c-library, and cannot be resolved in any solution depending on that library, hence the desire for a pure Python solution.
[…]
