### 2025-12-17 d2b2ab5
A few more changes after reading the formatted paper.
