### 2025-12-17 eda8fd0
switch to an MIT type NCAS-CMS license
