### 2025-12-17 aa9da56
update license info in pyproject toml
