+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
-    "setuptools >= 40.6.0",
-    "setuptools_scm>=6.2",
+    "setuptools>=80",
+    "setuptools_scm>=8",
 ]
 build-backend = "setuptools.build_meta"
 
@@ -40,7 +40,7 @@ dependencies = [
     "numpy>=2",
 ]
 description = "A pure python HDF5 reader"
-license = {text = "BSD License, Version 3-Clause"}
+license = {text = "MIT"}
 name = "pyfive"
 requires-python = ">=3.10"
 
