+++ .github/workflows/build-and-deploy-on-pypi.yml
@@ -7,6 +7,7 @@ on:
   push:
     branches:
       - main
+      - update_license_package
 
 jobs:
   build-n-publish:
