### 2025-12-17 #9630 @danielskatz
Thanks @bnlawrence - what you've done here is fine. Recommendations either on JOSS's list or outside are fine, and Github usernames (without the @) are preferred
