### 2025-12-18 8819431
add changelog for v1.1.0
