+++ Changelog.rst
@@ -1,4 +1,4 @@
-Version 1.1.0
+Version 1.0.1
 -------------
 
 **2025-12-18**
