### 2025-12-25 jjhelmus/pyfive#167 @dostuffthatmatters
[JOSS REVIEW]: Extend `.gitignore`
**This issue is part of the JOSS review of PyFive (https://github.com/openjournals/joss-reviews/issues/9688)**

Dear maintainers,

thank you for your work on PyFive! The package seems very useful and I don't think a lot has to change during the review.

Currently the `.gitignore` does not contain a lot of items. When I first created a venv in a `.venv` directory, it was not ignored. This should cover a reasonable set of files/directories on all major operating systems: https://www.toptal.com/developers/gitignore?templates=python,linux,macos,windows

Best,
Moritz Oliveira Makowski

_Doctoral Candidate
Environmental Sensing and Modeling (Prof. Jia Chen)
Technical University of Munich_
