### 2025-12-30 8830c73
negative slices/integers
