+++ pyfive/h5d.py
@@ -15,6 +15,7 @@ from importlib.metadata import version
 StoreInfo = namedtuple('StoreInfo',"chunk_offset filter_mask byte_offset size")
 ChunkIndex = namedtuple('ChunkIndex',"chunk_address chunk_dims")
 
+
 class DatasetID:
     """ 
     Implements an "HDF5 dataset identifier", which despite the name, actually
@@ -570,13 +571,16 @@ class DatasetID:
             fh.close()
 
         return out
-
+#ppp
     def _get_selection_via_chunks(self, args):
         """Use the zarr orthogonal indexer to extract data for a specfic
         selection within the dataset array and in doing so, only load
         the relevant chunks.
 
         """
+        args, squeeze_dims, flip_dims = self._parse_indices(args, self.shape)
+
+        print (args, squeeze_dims, flip_dims)
         # need a local dtype as we may override it for a reference read.
         dtype = self.dtype
         if isinstance(self._ptype, P5ReferenceType):
@@ -640,7 +644,141 @@ class DatasetID:
             to_reference = np.vectorize(Reference)
             out = to_reference(out)
 
+        if squeeze_dims:
+            # Squeeze dimensions that were indexed with an integer or
+            # 0-d numpy array.
+            out = np.squeeze(out, axis=squeeze_dims)
+
+        if flip_dims:
+            # Flip dimensions that were indexed with a negative-step
+            # slice
+            out = np.flip(out, axis=flip_dims)
+
         return out
+
+    def _parse_indices(self,indices, shape):
+        """Reformat the indices for assignment.
+
+        An int or scalar numpy array is recast as an a size 1, 1-d
+        numpy array. The dimensions in the output array for which this
+        has occurred are also returned so that they can be squeezed
+        later.
+        
+        A slice object that is decreasing (i.e. with a negative step),
+        is recast as an increasing slice (i.e. with a positive
+        step. For example ``slice(7,3,-1)`` would be cast as
+        ``slice(4,8,1)``. This is to facilitate finding which chunks
+        are touched by the index. The dimensions in the output array
+        for which this has occurred are also returned so that they can
+        be flipped later.
+
+        Parameters
+        ----------
+        indices : numpy-style indices
+            Indices to array defining the elements to be assigned.
+
+        shape : sequence of `int`
+            The shape of the array to be indexed.
+    
+        Returns
+        -------
+        modified_indices : `tuple`
+            The reformatted indices that give the correct subspace
+            after any output dimensions have been flipped or squeezed,
+            as appropriate.
+
+        squeeze_dims : `tuple`
+            The positions in the output array of any size 1 dimensions
+            that will need squeezing.
+        
+        flip_dims : `tuple`
+            The positions in the output array of any dimensions that
+            will need flipping.
+    
+        Examples
+        --------
+        >>> _parse_indices((slice(1, -1),), (8,))
+        ((slice(1, 7, 1),), [])
+    
+        >>> _parse_indices(([1, 2, 6, 5],), (8,))
+        ((array([1, 2, 6, 5]),), [])
+    
+        >>> _parse_indices((slice(-1, 2, -1), 3), (7, 8))
+        ((slice(3, 8, 1), 3), [0])
+    
+        >>> _parse_indices((3, slice(-1, 2, -1)), (7, 8))
+        ((3, slice(3, 8, 1)), [0])
+    
+        >>> _parse_indices(([3], slice(-1, 2, -1)), (7, 8))
+        (([3], slice(3, 8, 1)), [1])
+    
+        >>> _parse_indices(([1, 2], [3], slice(-1, 2, -1)), (7, 8, 9))
+        (([1, 2], [3], slice(3, 7, 1)), [2])
+
+        """
[…]
