+++ pyfive/h5d.py
@@ -187,37 +187,59 @@ class DatasetID:
                 if self._data is None:
                     no_storage = True
                 else:
-                    return self._read_compact_data(args)
+                    out = self._read_compact_data(args)
+
             case 1:  # contiguous storage
                 if self.data_offset == UNDEFINED_ADDRESS:
                     no_storage = True
                 else:
-                    return self._get_contiguous_data(args, fillvalue)
+                    out = self._get_contiguous_data(args, fillvalue)
+
             case 2:  # chunked storage
                 if not self.__index_built:
                     self._build_index()
+
                 if not self._index:
                     no_storage = True
                 else:
+                    args, squeeze_dims, flip_dims = (
+                        self._parse_indices_for_chunks(args)
+                    )
+
                     if isinstance(self._ptype, P5ReferenceType):
                         # references need to read all the chunks for now
-                        return self._get_selection_via_chunks(())[args]
+                        out = self._get_selection_via_chunks(())[args]
                     else:
                         # this is lazily reading only the chunks we need
-                        return self._get_selection_via_chunks(args)
+                        out = self._get_selection_via_chunks(args)
+
+                    if flip_dims:
+                        # Flip dimensions that were indexed with a
+                        # negative-step slice
+                        out = np.flip(out, axis=flip_dims)
+
+                    if squeeze_dims:
+                        # Squeeze dimensions that were indexed with an
+                        # integer or 0-d numpy array.
+                        out = np.squeeze(out, axis=squeeze_dims)
 
         if no_storage:
-            return np.full(self.shape, fillvalue, dtype=self.dtype)[args]
+            out = np.full(self.shape, fillvalue, dtype=self.dtype)[args]
 
+        return out
 
     def iter_chunks(self, args):
-        """ 
-        Iterate over chunks in a chunked dataset. 
-        The optional sel argument is a slice or tuple of slices that defines the region to be used. 
-        If not set, the entire dataspace will be used for the iterator.
-        For each chunk within the given region, the iterator yields a tuple of slices that gives the
-        intersection of the given chunk with the selection area. 
-        This can be used to read data in that chunk.
+        """Iterate over chunks in a chunked dataset.
+
+        The args argument is a (possibly empty) sequence of indices
+        that defines the region to be used. If an empty sequence then
+        the entire dataspace will be used for the iterator.
+
+        For each chunk within the given region, the iterator yields a
+        tuple of indices that gives the intersection of the given
+        chunk with the selection area. This can be used to read data
+        in that chunk.
+
         """
         if not self.__chunk_init_check():
             return None
@@ -382,7 +404,6 @@ class DatasetID:
         self.__index_built=True
 
     def _get_contiguous_data(self, args, fillvalue):
-
         if isinstance(self._ptype, P5ReferenceType):
             size = self._ptype.size
             if size != 8:
@@ -448,6 +469,7 @@ class DatasetID:
                         new_array[:] = result
                         result = _decode_array(result, new_array)
                 fh.close()
+                print (999)
                 return result
             except UnsupportedOperation:
                 return self._get_direct_from_contiguous(args)
@@ -571,16 +593,13 @@ class DatasetID:
             fh.close()
 
         return out
-#ppp
+
     def _get_selection_via_chunks(self, args):
         """Use the zarr orthogonal indexer to extract data for a specfic
         selection within the dataset array and in doing so, only load
         the relevant chunks.
 
         """
[…]
