+++ Changelog.rst
@@ -1,3 +1,10 @@
+Version ?.?.?
+-------------
+
+**2026-??-??**
+
+* Allow negative slices when indexing chunked data by `David Hassell <https://github.com/davidhassell>`_ in https://github.com/NCAS-CMS/pyfive/pull/???
+
 Version 1.0.1
 -------------
 
