### 2026-01-05 8a6be77
Allow negative slices when indexing chunked data
