+++ tests/test_dataset_indexing.py
@@ -8,7 +8,6 @@ import pytest
 import pyfive
 
 DIRNAME = os.path.dirname(__file__)
-DATASET_CONTIGUOUS_FILE = os.path.join(DIRNAME, 'data/dataset_multidim.hdf5')
 DATASET_CHUNKED_FILE = os.path.join(DIRNAME, 'data/chunked.hdf5')
 
 # Define a fixture that opens the chunked file once
@@ -37,24 +36,27 @@ def test_dataset_indexing_chunked_negative_step_slices(chunked_file, index):
 
     assert_array_equal(d[index], array[index])
 
-def test_dataset_indexing_parse_indices_for_chunks():
-    """Test pyfive.indexing.parse_indices_for_chunks"""
-    func = pyfive.indexing.parse_indices_for_chunks
+def test_dataset_indexing_replace_negative_slices():
+    """Test pyfive.indexing.replace_negative_slices"""
+    func = pyfive.indexing.replace_negative_slices
 
-    range8 = list(range(8))
-
-    assert func((Ellipsis,), (7, 8, 9)) == (
-        (slice(0, 7, 1), slice(0, 8, 1), slice(0, 9, 1)), ()
-    )
-
+    r = list(range(8))
     assert func((0, slice(6, 0, -2)), (7, 8, 9)) == (
-        (0, slice(2, 7, 2), slice(0, 9, 1)), (0,)
+        (0, slice(2, 7, 2)), (0,)
     )
-    assert range8[slice(6, 0, -2)] == list(reversed(range8[slice(2, 7, 2)]))
+    assert r[slice(6, 0, -2)] == list(reversed(r[slice(2, 7, 2)]))
 
-    assert func(([1, 2], slice(600, None, -3), ..., 0), (7, 8, 9)) == (
-        ([1, 2], slice(1, 8, 3), 0), (1,)
+    assert func(([1, 2], slice(600, 1, -3), 0), (7, 8, 9)) == (
+        ([1, 2], slice(4, 8, 3), 0), (1,)
     )
-    assert range8[slice(600, None, -3)] == list(reversed(range8[slice(1, 8, 3)]))
+    assert r[slice(600, None, -3)] == list(reversed(r[slice(1, 8, 3)]))
 
+    r = list(range(9))
+    assert func((0, slice(None), slice(6, 0, -2)), (7, 8, 9)) == (
+        (0, slice(None), slice(2, 7, 2)), (1,)
+    )
+    assert r[slice(6, 0, -2)] == list(reversed(r[slice(2, 7, 2)]))
 
+    # Can't pass in in Ellipsis
+    with pytest.raises(ValueError) as error:
+       func((0, slice(6, 0, -2), ...), (7, 8, 9))
