### 2026-01-05 d16a773
Add __orthogonal_indexing__ flag
