### 2026-01-05 NCAS-CMS/pyfive#168 @bnlawrence
@valeriupredoi Could you please think about this. I am not sure that putting static type hints in *everywhere* is practical on a review timescale, but we could and should consider providing them for all public _high-level_ API end points, and then add some sort of "static api coverage check" to our commit actions so we can pick up the rest over time. This probably plays with #121, which we could bring forward.
