### 2026-01-05 082b00f
improve variable names and comments
