### 2026-01-05 bacfb99
better Ellipsis error message
