### 2026-01-07 NCAS-CMS/pyfive#167 @kmuehlbauer
While I agree with @valeriupredoi's reasoning, it might still be worthwhile to think about better usage of `.gitignore`. As a minimal solution we could tweak `.gitignore` to act like **_Ignore everything, allow only what matters_**.
