### 2026-01-07 jjhelmus/pyfive#172 @Zeitsperre
[JOSS REVIEW] Synchronize development dependencies
Hey there,

(I'm starting my review on this and will be opening a few more small issues to address potential problems as I happen upon them.)

It appears that there's a disconnect between the dependencies listed in `environment.yml` and those in `pyproject.toml`. For most projects I've seen, these should more or less match (insofar as binaries are available on conda-forge). Easy to address.

Context: https://github.com/openjournals/joss-reviews/issues/9688
