### 2026-01-07 19bd58e
enable mypy in pyproject toml
