+++ pyproject.toml
@@ -113,3 +113,10 @@ exclude_lines = [
     "if __name__ == .__main__.:",
     "if TYPE_CHECKING:",
 ]
+
+[tool.mypy]
+# See https://mypy.readthedocs.io/en/stable/config_file.html
+ignore_missing_imports = true
+enable_error_code = [
+    "truthy-bool",
+]
