### 2026-01-07 4c9757b
add pre commit config
