### 2026-01-07 jjhelmus/pyfive#176 @Zeitsperre
[JOSS REVIEW] Adopt PEP 639 for license
The license information detailed in the `pyproject.toml` is inaccurate:

> `license = {text = "BSD License, Version 3-Clause"},`

The license should make use of the SPDX code using the PEP 639 standard (supported by `setuptools >=77.0`):

> `license = "BSD-3-Clause",`
> `license-files = ["LICENSE.txt"],`

At the same time, the license classifier (`"License :: OSI Approved :: BSD License",`) can be removed (now deprecated).

Resources: https://peps.python.org/pep-0639/

Context: https://github.com/openjournals/joss-reviews/issues/9688
