### 2026-01-07 NCAS-CMS/pyfive#177 @Zeitsperre
[JOSS REVIEW] Add CONTRIBUTING guidelines
It would be great to have a dedicate section for advice on how best to contribute to the project. Some examples of relevant information:

- How to fork and clone the repo
- Suggested development installation steps
- Types of contributions that are accepted
- How to run the tests / build the docs
- Quality control steps
- Other considerations for Pull Requests

This is a relatively generic (albeit comprehensive) example from a repo I maintain: https://github.com/Ouranosinc/xsdba/blob/main/CONTRIBUTING.rst. Feel free to be inspired/copy the relevant portions as needed. This could also be wrapped into the discussion for #150.

Context: https://github.com/openjournals/joss-reviews/issues/9688
