### 2026-01-07 NCAS-CMS/pyfive#178 @Zeitsperre
[JOSS REVIEW] Specify linting rules in a config file
In the workflows, there are linting calls that have flagged values (for `flake8`):
```yaml
- name: Lint with flake8
  run: |
    # stop the build if there are Python syntax errors or undefined names
    flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
    # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
    flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
```

It would be better to have these `flake8` (or `ruff`) rules specified in a configuration file (`flake8`: `.flake8`/`setup.cfg`, `ruff`: `pyproject.toml`). The default values expected for cyclomatic complexity and line length can also be adjusted (for both projects).

Related: #121

Context: https://github.com/openjournals/joss-reviews/issues/9688
