### 2026-01-07 0c9dea0
small formatting nitpicks
Signed-off-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
