+++ doc/quickstart/installation.rst
@@ -63,6 +63,7 @@ We adhere to `Scientific Python SPEC-0 <https://scientific-python.org/specs/spec
 * 3.11
 * 3.12
 * 3.13
+* 3.14
 
 .. note::
 
