### 2026-01-08 ebf3464
add pre-commit test dependency
