+++ environment.yml
@@ -17,6 +17,7 @@ dependencies:
   - h5py  # link to same hdf5 lib with netcdf4
   - moto
   - netcdf4  # has history of broken packages on PyPI
+  - pre-commit
   - pytest
   - pytest-cov
   - pytest-html
