### 2026-01-08 e00bd39
run pre-commit in GHA CI
