+++ .github/workflows/pytest.yml
@@ -45,6 +45,10 @@ jobs:
           pip install -e .[test]
       - name: Run pip check
         run: pip check
+      - name: Run pre-commit with mypy
+        run: |
+          pre-commit install
+          pre-commit run -a
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
