+++ .github/workflows/pytest.yml
@@ -49,6 +49,7 @@ jobs:
         run: |
           pre-commit install
           pre-commit run -a
+      # this will be removed when fully Ruff-ed
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
