### 2026-01-08 ea101f2
add ruff rulesets in pyrpoject toml
