+++ pyproject.toml
@@ -107,13 +107,70 @@ addopts = [
 [tool.coverage.run]
 parallel = true
 source = ["pyfive"]
-    
+
 [tool.coverage.report]
 exclude_lines = [
     "pragma: no cover",
     "if __name__ == .__main__.:",
     "if TYPE_CHECKING:",
 ]
+
+[tool.ruff]
+line-length = 100
+[tool.ruff.lint]
+select = ["ALL"]
+ignore = [
+    # Ignore until someone has time to fix this:
+    "DTZ",      # add tzinfo to datetimes
+    "FBT",      # avoid boolean positional arguments
+    "N818",     # Exception name should be named with an Error suffix
+    "PERF",     # Use comprehensions
+    "PGH003",   # Use specific rule codes when ignoring type issues
+    "PLR2004",  # Don't use magic values
+    "PT009",    # Use a regular `assert` instead of unittest-style `assertListEqual`
+    "PT011",    # pytest.raises without match argument
+    "PT027",    # Use `pytest.raises` instead of unittest-style `assertRaises`
+    "PTH",      # Use pathlib everywhere
+    "SIM102",   # Use a single `if` statement instead of nested `if` statements
+    "SIM117",   # Use a single `with` statement with multiple contexts instead of nested `with` statements
+    "TD003",    # Missing issue link on the line following this TODO
+    "TRY400",   # Uses of `logging.error` instead of `logging.exception` when logging an exception
+    # Really ignore:
+    "F401",     # Need to keep unused imports (from pyfive.h5d import DatasetID)
+    "D105",     # Disable missing docstring in magic method as these are well defined.
+    "E501",     # Disable line-too-long as this is taken care of by the formatter.
+    "FIX002",   # TODOs can be useful when something cannot be done immediately.
+    "PLW1641",  # Implementing __hash__ on mutable objects may be more confusing than not doing it.
+    "S603",     # Uncheckable
+    "TD002",    # Missing author in TODO: can be seen from git history
+    "TRY003",   # Fixing this would require many exceptions
+]
+future-annotations = true
+[tool.ruff.lint.per-file-ignores]
+"tests/**.py" = [
+    "ARG",      # Fixtures that set up some state are wrongly identified as unused arguments.
+    "S101",     # Use of assert in tests code is required.
+    "T20",      # Printing things can help with debugging tests.
+    "SLF001",   # Testing private members is ok
+    "PLR2004",  # Magic values in tests are fine.
+    # Docstrings in tests are only needed if the code is not self-explanatory.
+    "D100",     # Missing docstring in public module
+    "D101",     # Missing docstring in public class
+    "D102",     # Missing docstring in public method
+    "D103",     # Missing docstring in public function
+    "D104",     # Missing docstring in public package
+    "PT013",    # Allow importing fixtures from pytest to avoid repeating 'pytest' many times
+]
+[tool.ruff.lint.flake8-annotations]
+allow-star-arg-any = true
+ignore-fully-untyped = true
+# turn these on at a later stage
+# [tool.ruff.lint.isort]
+# known-first-party = ["pyfive"]
+# [tool.ruff.lint.pydocstyle]
+# convention = "numpy"
+# [tool.ruff.lint.pylint]
+# max-args = 7
 
 [tool.mypy]
 # See https://mypy.readthedocs.io/en/stable/config_file.html
