### 2026-01-08 8ee4356
add rule
