### 2026-01-08 7bc1a04
restrict checks only to pyfive for now
