+++ pyproject.toml
@@ -136,6 +136,7 @@ ignore = [
     # Really ignore:
     "F401",     # Need to keep unused imports (from pyfive.h5d import DatasetID)
     "D105",     # Disable missing docstring in magic method as these are well defined.
+    "D107",     # Disable Missing docstring in `__init__`
     "E501",     # Disable line-too-long as this is taken care of by the formatter.
     "FIX002",   # TODOs can be useful when something cannot be done immediately.
     "PLW1641",  # Implementing __hash__ on mutable objects may be more confusing than not doing it.
@@ -145,6 +146,8 @@ ignore = [
     "D400",     # First line should end with a period
     "D401",
     "D404",     # three superficial complaints about docstrings (imperative, punctuation etc)
+    "S101",     # Use of `assert` detected
+    "SLF001",   # Testing private members is ok
     "T201",     # `print` found - leave it
 ]
 future-annotations = true
