### 2026-01-08 42aca09
Apply suggestions from code review
Co-authored-by: David Hassell <davidhassell@users.noreply.github.com>
