### 2026-01-08 6e09843
add rule
