### 2026-01-08 de65c74
turn off mypy temporarily
