+++ pyproject.toml
@@ -164,9 +164,10 @@ ignore-fully-untyped = true
 # [tool.ruff.lint.pylint]
 # max-args = 7
 
-[tool.mypy]
+# turn off temporarily
+#[tool.mypy]
 # See https://mypy.readthedocs.io/en/stable/config_file.html
-ignore_missing_imports = true
-enable_error_code = [
-    "truthy-bool",
-]
+#ignore_missing_imports = true
+#enable_error_code = [
+#    "truthy-bool",
+#]
