+++ .pre-commit-config.yaml
@@ -36,11 +36,11 @@ repos:
       - id: ruff-check
         args: [--fix]
       - id: ruff-format
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.19.1'
-    hooks:
-      - id: mypy
-        additional_dependencies:
-          - 'types-PyYAML'
-          - 'types-requests'
-          - 'numpy'
+#  - repo: https://github.com/pre-commit/mirrors-mypy
+#    rev: 'v1.19.1'
+#    hooks:
+#      - id: mypy
+#        additional_dependencies:
+#          - 'types-PyYAML'
+#          - 'types-requests'
+#          - 'numpy'
