### 2026-01-09 9c9eaef
move to a more restrictive ruff linting
