### 2026-01-09 d73cc14
readd tests to analysis
