+++ .pre-commit-config.yaml
@@ -6,9 +6,10 @@ ci:
 
 exclude: |
   (?x)
+  .readthedocs.yaml|
+  codecov.yml|
   ^.github/|
-  ^doc/|
-  ^tests/
+  ^doc/
 
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
