### 2026-01-09 036c645
correct line length setter
