+++ .pre-commit-config.yaml
@@ -27,11 +27,12 @@ repos:
     rev: 'v1.37.1'
     hooks:
       - id: yamllint
-  - repo: https://github.com/codespell-project/codespell
-    rev: 'v2.4.1'
-    hooks:
-      - id: codespell
-        additional_dependencies: [tomli]  # required for Python 3.10
+  # codespell is always iffy; we can turn it on if we want to
+  # - repo: https://github.com/codespell-project/codespell
+  #   rev: 'v2.4.1'
+  #   hooks:
+  #     - id: codespell
+  #       additional_dependencies: [tomli]  # required for Python 3.10
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: "v0.14.9"
     hooks:
