+++ pyfive/misc_low_level.py
@@ -1,4 +1,4 @@
-""" Misc low-level representation of HDF5 objects. """
+"""Misc low-level representation of HDF5 objects."""
 
 import struct
 from math import log2
@@ -16,7 +16,7 @@ from math import prod
 import numpy as np
 
 # uncomment this and use as shown in the FractalHeap if I/O diagnostic is needed
-#from .utilities import Interceptor
+# from .utilities import Interceptor
 
 
 class SuperBlock(object):
@@ -25,10 +25,10 @@ class SuperBlock(object):
     """
 
     def __init__(self, fh, offset):
-        """ initalize. """
+        """initalize."""
 
         fh.seek(offset)
-        version_hint = struct.unpack_from('<B', fh.read(9), 8)[0]
+        version_hint = struct.unpack_from("<B", fh.read(9), 8)[0]
         fh.seek(offset)
         if version_hint == 0:
             contents = _unpack_struct_from_file(SUPERBLOCK_V0, fh)
@@ -38,27 +38,28 @@ class SuperBlock(object):
             self._end_of_sblock = offset + SUPERBLOCK_V2_V3_SIZE
         else:
             raise NotImplementedError(
-                "unsupported superblock version: %i" % (version_hint))
+                "unsupported superblock version: %i" % (version_hint)
+            )
 
         # verify contents
-        if contents['format_signature'] != FORMAT_SIGNATURE:
-            raise InvalidHDF5File('Incorrect file signature')
-        if contents['offset_size'] != 8 or contents['length_size'] != 8:
-            raise NotImplementedError('File uses none 64-bit addressing')
-        self.version = contents['superblock_version']
+        if contents["format_signature"] != FORMAT_SIGNATURE:
+            raise InvalidHDF5File("Incorrect file signature")
+        if contents["offset_size"] != 8 or contents["length_size"] != 8:
+            raise NotImplementedError("File uses none 64-bit addressing")
+        self.version = contents["superblock_version"]
         self._contents = contents
         self._root_symbol_table = None
         self._fh = fh
 
     @property
     def offset_to_dataobjects(self):
-        """ The offset to the data objects collection for the superblock. """
+        """The offset to the data objects collection for the superblock."""
         if self.version == 0:
             sym_table = SymbolTable(self._fh, self._end_of_sblock, root=True)
             self._root_symbol_table = sym_table
             return sym_table.group_offset
         elif self.version == 2 or self.version == 3:
-            return self._contents['root_group_address']
+            return self._contents["root_group_address"]
         else:
             raise NotImplementedError
 
@@ -69,21 +70,21 @@ class Heap(object):
     """
 
     def __init__(self, fh, offset):
-        """ initalize. """
+        """initalize."""
 
         fh.seek(offset)
         local_heap = _unpack_struct_from_file(LOCAL_HEAP, fh)
-        assert local_heap['signature'] == b'HEAP'
-        assert local_heap['version'] == 0
-        fh.seek(local_heap['address_of_data_segment'])
-        heap_data = fh.read(local_heap['data_segment_size'])
-        local_heap['heap_data'] = heap_data
+        assert local_heap["signature"] == b"HEAP"
+        assert local_heap["version"] == 0
+        fh.seek(local_heap["address_of_data_segment"])
+        heap_data = fh.read(local_heap["data_segment_size"])
+        local_heap["heap_data"] = heap_data
         self._contents = local_heap
         self.data = heap_data
 
     def get_object_name(self, offset):
-        """ Return the name of the object indicated by the given offset. """
-        end = self.data.index(b'\x00', offset)
+        """Return the name of the object indicated by the given offset."""
+        end = self.data.index(b"\x00", offset)
[…]
