+++ pyfive/high_level.py
@@ -1,4 +1,4 @@
-""" High-level classes for reading HDF5 files.  """
+"""High-level classes for reading HDF5 files."""
 
 from collections import deque
 from collections.abc import Mapping, Sequence
@@ -32,7 +32,7 @@ class Group(Mapping):
     """
 
     def __init__(self, name, dataobjects, parent):
-        """ initalize. """
+        """initalize."""
 
         self.parent = parent
         self.file = parent.file
@@ -46,30 +46,29 @@ class Group(Mapping):
         return '<HDF5 group "%s" (%d members)>' % (self.name, len(self))
 
     def __len__(self):
-        """ Number of links in the group. """
+        """Number of links in the group."""
         return len(self._links)
 
     def _dereference(self, ref):
-        """ Dereference a Reference object. """
+        """Dereference a Reference object."""
         if not ref:
-            raise ValueError('cannot deference null reference')
+            raise ValueError("cannot deference null reference")
         obj = self.file._get_object_by_address(ref.address_of_reference)
         if obj is None:
-            raise ValueError('reference not found in file')
+            raise ValueError("reference not found in file")
         return obj
 
     def __getitem__(self, y):
-        """ x.__getitem__(y) <==> x[y].
-        """
+        """x.__getitem__(y) <==> x[y]."""
         return self.__getitem_lazy_control(y, noindex=False)
 
     def get_lazy_view(self, y):
-        """ 
-        This instantiates the object y, and if it is a 
+        """
+        This instantiates the object y, and if it is a
         chunked dataset, does so without reading the b-tree
         index. This is useful for inspecting a variable
         that you are not expecting to access. If you know you
-        want to access the data, and in particular, if you are 
+        want to access the data, and in particular, if you are
         going to hand the data to Dask or something else, you
         almost certainly want to read the index now, so
         just do x[y] rather than x.get_lazy_view(y).
@@ -80,18 +79,18 @@ class Group(Mapping):
         return self.__getitem_lazy_control(y, noindex=True)
 
     def __getitem_lazy_control(self, y, noindex):
-        """ 
+        """
         This is the routine which actually does the get item
         but does it in such a way that we control how much laziness
         is possible where we have chunked variables with b-trees.
 
         We want to return y, but if y is a chunked dataset we
         normally return it with a cached b-tree (noindex=false).
-        If noindex is True, we do not read the b-tree, and that 
+        If noindex is True, we do not read the b-tree, and that
         will be done when data is first read - which is fine
         in a single-threaded environment, but in a parallel
         environment you only want to read the index once
-        (so use noindex=False, which you get via the 
+        (so use noindex=False, which you get via the
         normal getitem interface - x[y]).
         """
 
@@ -99,19 +98,19 @@ class Group(Mapping):
             return self._dereference(y)
 
         path = posixpath.normpath(y)
-        if path == '.':
+        if path == ".":
             return self
-        if path.startswith('/'):
+        if path.startswith("/"):
             return self.file[path[1:]]
 
-        if posixpath.dirname(path) != '':
-            next_obj, additional_obj = path.split('/', 1)
+        if posixpath.dirname(path) != "":
+            next_obj, additional_obj = path.split("/", 1)
         else:
             next_obj = path
-            additional_obj = '.'
+            additional_obj = "."
 
         if next_obj not in self._links:
-            raise KeyError('%s not found in group' % (next_obj))
+            raise KeyError("%s not found in group" % (next_obj))
 
[…]

[…]
