+++ tests/test_fractal_heap.py
@@ -4,7 +4,8 @@ import pytest
 from contextlib import nullcontext
 import pyfive
 
-@pytest.fixture(scope='module')
+
+@pytest.fixture(scope="module")
 def name(tmp_path_factory):
     return tmp_path_factory.mktemp("temp") / "fractal_heap.hdf5"
 
@@ -23,7 +24,9 @@ def test_huge_object(name, payload_size, n_attrs):
 
     with h5py.File(name, "w", track_order=True) as f:
         for i in range(n_attrs):
-            f.attrs[f"small_{i}"] = np.random.randint(low=0, high=255, size=payload_size, dtype=np.uint8)
+            f.attrs[f"small_{i}"] = np.random.randint(
+                low=0, high=255, size=payload_size, dtype=np.uint8
+            )
 
     with h5py.File(name, "r") as f:
         attrs = dict(f.attrs)
@@ -36,15 +39,14 @@ def test_huge_object(name, payload_size, n_attrs):
             for k, v in attrs.items():
                 np.testing.assert_equal(v, attrs2[k])
 
+
 @pytest.mark.parametrize("n_attrs", [115, 116])
 def test_fractal_heap(name, n_attrs):
-
     # att: the assumptions below might heavily rely on the
     # file layout, heaps sizes and other figures
     # todo: generalize this
 
     with h5py.File(name, "w", track_order=True) as f:
-
         # create enough attributes to trigger dense storage
         # and indirect blocks
         # using small payloads to control the block filling
@@ -55,7 +57,9 @@ def test_fractal_heap(name, n_attrs):
         # from 4033 this will run into huge object space
         payload_size = 4032
         for i in range(n_attrs):
-            f.attrs[f"attr_{i}"] = np.random.randint(low=0, high=255, size=payload_size, dtype=np.uint8)
+            f.attrs[f"attr_{i}"] = np.random.randint(
+                low=0, high=255, size=payload_size, dtype=np.uint8
+            )
 
     with h5py.File(name, "r") as f:
         attrs = dict(f.attrs)
@@ -64,9 +68,11 @@ def test_fractal_heap(name, n_attrs):
         print("\n--- debug output for test -----------------------\n")
         # since we can't get any information on the heap object from pyfive
         attr_info = f._dataobjects.find_msg_type(0x0015)
-        offset = attr_info[0]['offset_to_message']
-        data = pyfive.core._unpack_struct_from(pyfive.dataobjects.ATTR_INFO_MESSAGE, f._dataobjects.msg_data, offset)
-        heap_address = data['fractal_heap_address']
+        offset = attr_info[0]["offset_to_message"]
+        data = pyfive.core._unpack_struct_from(
+            pyfive.dataobjects.ATTR_INFO_MESSAGE, f._dataobjects.msg_data, offset
+        )
+        heap_address = data["fractal_heap_address"]
         heap = pyfive.misc_low_level.FractalHeap(f._fh, heap_address)
 
         # nfortunately we can't get anything meaningful out of this
