### 2026-01-09 8f79cea
turn mypy back on
