+++ tests/test_high_level.py
@@ -13,7 +13,7 @@ EARLIEST_HDF5_FILE = os.path.join(DIRNAME, "data", "earliest.hdf5")
 
 # Polyglot string type for representing unicode
 try:
-    string_type = unicode
+    string_type = unicode  # type: ignore[name-defined]
 except NameError:
     string_type = str
 
