### 2026-01-09 ca71d9c
turn off some mypy typing bits
