### 2026-01-09 7d11443
started added type hints
