### 2026-01-13 3fc9141
type hints
