### 2026-01-13 cf2f848
more type hinting
