### 2026-01-13 862936f
add a missing type return
