### 2026-01-13 20fe4a2
keep generic {} but type the var
