+++ pyfive/h5d.py
@@ -98,7 +98,7 @@ class DatasetID:
         # each variable to have it's own copy of those needed for
         # data access. Though that's clearly not optimal if they include
         # other data. To be determined.
-        self._global_heaps = dict
+        self._global_heaps: dict = {}
 
         self._msg_offset, self.layout_class, self.property_offset = (
             dataobject.get_id_storage_params()
