### 2026-01-14 4ae10b1
add str type
