### 2026-01-14 5486e7f
add return and ignore type missing return - inconsistent returns in func but they should not be changed, this is a common mypy false negative
