### 2026-01-14 dbbb925
turn on mypy section
