+++ pyproject.toml
@@ -57,7 +57,7 @@ test = [
     "h5py",
     "moto",
     "netCDF4",
-    "pre-commit",
+    "pre-commit",  # with ruff and mypy
     "pytest",
     "pytest-cov",
     "pytest-html",
@@ -147,10 +147,9 @@ ignore-fully-untyped = true
 # [tool.ruff.lint.pylint]
 # max-args = 7
 
-# turn off temporarily
-#[tool.mypy]
+[tool.mypy]
 # See https://mypy.readthedocs.io/en/stable/config_file.html
-#ignore_missing_imports = true
-#enable_error_code = [
-#    "truthy-bool",
-#]
+ignore_missing_imports = true
+enable_error_code = [
+    "truthy-bool",
+]
