+++ environment.yml
@@ -10,7 +10,6 @@ dependencies:
   - python >=3.10
   # testing dependencies
   - dask
-  - flake8
   - flask
   - flask-cors
   - h5netcdf
