+++ environment.yml
@@ -15,7 +15,7 @@ dependencies:
   - h5netcdf
   - h5py  # link to same hdf5 lib with netcdf4
   - moto
-  - netcdf4  # has history of broken packages on PyPI
+  - netcdf4 !=1.7.4  # github.com/Unidata/netcdf4-python/issues/1464
   - pre-commit  # with ruff and mypy
   - pytest
   - pytest-cov
