+++ pyproject.toml
@@ -55,7 +55,7 @@ test = [
     "h5netcdf",
     "h5py",
     "moto",
-    "netCDF4",
+    "netCDF4!=1.7.4",  # github.com/Unidata/netcdf4-python/issues/1464
     "pre-commit",  # with ruff and mypy
     "pytest",
     "pytest-cov",
