### 2026-01-14 732cb57
use legacy import for Python 3.10-compat
