### 2026-01-14 defda34
modernize setuptools section
