### 2026-01-14 NCAS-CMS/pyfive#178 @Zeitsperre
That's precisely what I had in mind. Looking much better!
