+++ pyfive/high_level.py
@@ -513,20 +513,25 @@ class Dataset(object):
 
         """
         if self.id.chunks is not None:
-            # Chunked data
+            # Chunked data indexed with
+            # `DatasetID._get_selection_via_chunks`
             return True
 
-        if not (isinstance(self.id._ptype, (P5ReferenceType, P5VlenStringType, P5SequenceType))) and not self.id.posix:
-            # Contiguous data accessed with pseudo chunking via
-            # `DatasetID._get_direct_from_contiguous`.
-            #
-            # This test is based on the logic of
-            # `DatasetID._get_contiguous_data`, which decides whether
-            # or not to call `DatasetID._get_direct_from_contiguous`.
+        if (
+            not (
+                isinstance(
+                    self.id._ptype, (P5ReferenceType, P5VlenStringType, P5SequenceType)
+                )
+            )
+            and not self.id.posix
+        ):
+            # Contiguous data indexed with
+            # `DatasetID._get_direct_from_contiguous`
             return True
 
         # All other cases
         return False
+
 
 class DimensionManager(Sequence):
     """Represents a collection of dimensions associated with a dataset."""
