### 2026-01-26 #9688 @rich2355
Hi @Zeitsperre, @ayenpure, @dostuffthatmatters, I am just checking in to see the reviewing status. Please feel free to let me know if there is anything that can benefit from me!

Please correct me if wrong: I found that the author @bnlawrence seems to have addressed the questions @Zeitsperre raised, and is working on what @dostuffthatmatters asked.

@ayenpure what about your status?

Thank you for your time on this!

Best wishes,
RL
