### 2026-01-26 #9688 @Zeitsperre
@rich2355 Thanks for the check-in!

I see that many of my comments have been addressed by the authors. I'll have another chance to run over the library sometime this week. There are likely a few more things to address in the paper, but the code/documentation of the tool itself is looking better IMO.
