### 2026-01-26 jjhelmus/pyfive#183 @dostuffthatmatters
[JOSS REVIEW]: Tiny changes to manuscript and docs
**This issue is part of the JOSS review of PyFive (https://github.com/openjournals/joss-reviews/issues/9688)**

Dear authors of pyfive,

I agree with your arguments in the manuscript and nothing major is missing from it. Below you can find a few things I would update.

**Manuscript:**

- L11: "pyfive (https://pyfive.readthedo...)" I would avoid the link to the docs here in case you should ever change it (e.g., to an NCAS domain). The paper has the link to the repository on the left column.
- L12: "While it is not a complete implementation of all the specifications and capabilities of HDF5, it includes all the core functionality necessary to read gridded datasets, whether stored contiguously or with chunks." This statement is a bit vague. Could you maybe elaborate a bit, which features of HDF5 pyfive does not support. I would mention compression as one of the supported features. Does pyfive support all kinds of compression HDF5 supports? If this list of supported and unsupported features gets too long, you could also reference a docs page.
- L19: I think you should cite https://github.com/h5py/h5py. Their repo or docs do not have a citation guideline, but you can just cite the GitHub repo with its latest version.
- L27: "While newer formats are starting to get mindshare" it would be good to name and cite these never alternatives.
- L52: "Current Status" is a bit unspecific. Maybe rename it to something like "Current State of pyfive" or "Current State of The Project. Because "Current Status" could also mean state-of-the-art (without pyfive)
- L65: "To improve internal metadata access,pyfive" -> no space between "access" and "pyfive"

**Docs:**

- It is hard to get from the readthedocs page to GitHub. So maybe just add a link with the GitHub repo on top of the introduction page, or somewhere in the sidebar as a link.

Best,
Moritz Oliveira Makowski
