### 2026-01-26 NCAS-CMS/pyfive#184 @dostuffthatmatters
[JOSS REVIEW]: Document/showcase "levels of laziness" in docs
**This issue is part of the JOSS review of PyFive (https://github.com/openjournals/joss-reviews/issues/9688)**

Dear authors of pyfive,

given that your manuscript mentions the "several levels of laziness" (L65-L68), it would be nice to document them. From the API reference alone, I cannot easily tell how the different levels work or how to use them.

Best,
Moritz Oliveira Makowski
