### 2026-01-26 #9688 @dostuffthatmatters
Hi @rich2355

Thank you for checking in!

I almost finished my review checklist as well. The tool works well and the manuscript is pretty much complete. There are only a few minor things left from my side:

- [x] https://github.com/NCAS-CMS/pyfive/issues/167
- [x] https://github.com/NCAS-CMS/pyfive/issues/168
- [x] https://github.com/NCAS-CMS/pyfive/issues/184
- [x] https://github.com/NCAS-CMS/pyfive/issues/183

The only thing that might be worth to discuss are the static type hints on the user-facing API ([#168](https://github.com/NCAS-CMS/pyfive/issues/168)).
