### 2026-01-27 #9688 @ayenpure
@rich2355 sorry for the radio silence.

Before the new year I had a chance to go over the repository and I checked for general practices for good software development with code organization, CI, and documentation. I found everything satisfactory. The documentation also points to usage of the tool properly, which I tried quickly and they seemed to work fine. I've checked all my boxes and it's an all clear from me :)
Thanks for inviting me to the review and please let me know if I missed anything or can offer any additional help.
