### 2026-01-28 NCAS-CMS/pyfive#184 @bnlawrence
This is documented here: https://pyfive.readthedocs.io/en/latest/additional.html as part of the additional API support.
