### 2026-02-02 7bf88a7
A couple of fixes identified by Ruff after the merge
