### 2026-02-02 723c2ce
Paper and documentation changes to address #183
