### 2026-02-02 #9688 @dostuffthatmatters
Hi @rich2355,

All of my points are resolved as well. I highly recommend acceptance.
