### 2026-02-02 9454569
exclude paper files form pre-commit checks
