### 2026-02-02 a916c4b
one more paper-related exclusion
