### 2026-02-02 06c3261
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/adrienverge/yamllint: v1.37.1 → v1.38.0](https://github.com/adrienverge/yamllint/compare/v1.37.1...v1.38.0)
- [github.com/astral-sh/ruff-pre-commit: v0.14.9 → v0.14.14](https://github.com/astral-sh/ruff-pre-commit/compare/v0.14.9...v0.14.14)
