### 2026-02-02 #9688 @Zeitsperre
I've left comments in a review of the paper: https://github.com/NCAS-CMS/pyfive/pull/95#pullrequestreview-3741657312

Once those have been resolved, the authors have my recommendation for publishing. Great work!
