### 2026-02-03 6300c20
Changes in response to @zeitsperre's comments on #95.
