+++ paper.md
@@ -59,29 +59,29 @@ bibliography: paper.bib
 
 # Summary
 
-`pyfive` is an open-source thread-safe pure Python package for reading data stored in HDF5. 
+`pyfive` is an open-source and thread-safe pure Python package for reading data stored in HDF5. 
 While it is not a complete implementation of all the specifications and capabilities of HDF5, it includes all the core functionality necessary to read gridded datasets, whether stored contiguously or with chunks (with or without standard compression options).
-All data access is fully lazy, the data is only read from storage when the numpy data arrays are manipulated. Originally developed some years ago, the package has recently been upgraded to support lazy access, and to add missing features necessary for handling all the environmental data known to the authors. 
-It is now a realistic option for production data access in environmental science and more widely. 
+All data access is fully lazy as the data is only read from storage when the numpy data arrays are manipulated. Originally developed some years ago, the package has recently been expanded to support lazy data access, and to add missing features necessary for handling all the HDF5-based environmental data known to the authors. 
+It is now a realistic option for production data access in environmental science and more broadly across otehr domains.
 The API is based on that of `h5py` ([https://github.com/h5py/h5py](https://github.com/h5py/h5py), a Python shimmy over the HDF5 C-library which itself is not thread-safe), with some API extensions to help optimise remote access. 
 With these extensions, coupled with thread safety, many of the limitations precluding the efficient use of HDF5 (and netCDF4) on cloud storage have been removed.
 
 # Statement of need
 
-HDF5[^1][@FolEA11] is probably the most important data format in physical science.
-It is particularly important in environmental science given the fact that netCDF4[^2][@Rew06] is HDF5 under the hood. 
+HDF5[^1][@FolEA11] is arguably the most important data format in physical science.
+It is of particular importance in the environmental sciences which rely on the netCDF4[^2][@Rew06] data format, which itself uses the HDF data format underneath. 
 From satellite missions, to climate models and radar systems, the default binary format has been HDF5 for decades. 
-While newer formats are starting to get mindshare, there are petabytes, if not exabytes, of existing HDF5, and there are still many good use-cases for creating new data in HDF5. 
-However, despite the history, there are few libraries for reading HDF5 file data that do not depend on the official HDF5 library maintained by the HDF Group, and in particular, apart from `pyfive`, in Python there are none that cover the needs of environmental science. 
+While newer data formats are starting to get mindshare, there are petabytes, if not exabytes, of existing HDF5, and there remain many good use-cases for creating new data in the HDF5 format today. 
+However, despite its historical importance, there are few libraries available for reading HDF5 file data that do not depend on the official HDF5 library maintained by the HDF Group. In particular, apart from `pyfive`, there are no Python HDF5 libraries that address the data access needs of environmental science. 
 While the HDF5 c-library is reliable and performant, and battle-tested over decades, there are some caveats to depending upon it: 
-Firstly, it is not thread-safe,  secondly, the code is large and complex, and should anything happen to the financial stability of the HDF5 Group, it is not obvious the C-code could be maintained. 
+Firstly, it is not thread-safe. Secondly, the underlying code is large and complex, and should anything happen to the financial stability of the HDF5 Group, it is not obvious it could be maintained. 
 Finally, the code complexity also meant that it is not suitable for developing bespoke code for data recovery in the case of partially corrupt data. 
-From a long-term curation perspective both of these last two constraints are a concern. 
+From a long-term curation perspective these last two constraints present a major concern. 
 
-The issues of the dependency on a complex code maintained by one private company in the context of maintaining data access (over decades, and potentially centuries), can only be mitigated by ensuring that the data format is well documented, that data writers use only the documented features, and that public code exists which can be relatively easily maintained. 
[…]
