### 2026-02-03 5f25539
replace python-lzf package with lz4 package
