+++ pyproject.toml
@@ -54,6 +54,7 @@ test = [
     "flask-cors",
     "h5netcdf",
     "h5py",
+    "lz4",
     "moto",
     "netCDF4!=1.7.4",  # github.com/Unidata/netcdf4-python/issues/1464
     "pre-commit",  # with ruff and mypy
@@ -62,7 +63,6 @@ test = [
     "pytest-html",
     "pytest-rerunfailures",
     "pytest-xdist",
-    "python-lzf",
     "s3fs>=2025.9.0",  # v-mismatch with boto3 results in ancient s3fs
 ]
 
