### 2026-02-03 07cad8c
decompress with lz4 instead of python-lzf
