### 2026-02-04 07f8676
Added a test for logging, and fixed some pylance type complaints.
