### 2026-02-04 2d8eb30
Fixed a bug, a couple of pylance complaints, and added some logging
