### 2026-02-04 eb56a94
Fixed a subtle issue with iter_chunks arguments, and added some logging.
