+++ pyfive/dataobjects.py
@@ -7,6 +7,8 @@ import struct
 import warnings
 
 import numpy as np
+from time import time
+import logging
 
 from pyfive.datatype_msg import DatatypeMessage
 from pyfive.core import _padded_size, _structure_size
@@ -161,6 +163,7 @@ class DataObjects(object):
 
     def get_attributes(self):
         """Return a dictionary of all attributes."""
+        t0 = time()
         attrs = {}
         attr_msgs = self.find_msg_type(ATTRIBUTE_MSG_TYPE)
         for msg in attr_msgs:
@@ -174,6 +177,9 @@ class DataObjects(object):
         if attr_info:
             more_attrs = self._get_attributes_from_attr_info(attrs, attr_info)
             attrs.update(more_attrs)
+        t1 = time()-t0
+        logging.info('Obtained {len(attrs)} attributes, operation took {t1:.3}s {version("pyfive")}')
+        
         return attrs
 
     def _get_attributes_from_attr_info(self, attrs, attr_info):
