### 2026-02-04 1fa0283
Woops. f-string
