### 2026-02-04 688321e
More log/timing info
