### 2026-02-04 a43cd2f
Still adding logging
