### 2026-02-04 690f675
Still refining logging
