### 2026-02-04 #9688 @Zeitsperre
@rich2355 Not a silly question at all.

The changes I proposed are simply to correct a few last typos I found in the paper. Once those are merged/corrected, that will conclude my review. The paper looks good, and I'm excited to see the project get some recognition!
