### 2026-02-04 #9688 @rich2355
Thanks for the rapid reply!

@bnlawrence What I really need now would be your confirmation that those minor typos are fixed.
