### 2026-02-05 2c25fdb
Update paper.md to get the HDF5 group name right
Co-authored-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
