### 2026-02-05 9bba33e
We're chasing down why s3fs is not caching.
