### 2026-02-05 b8e55f0
Caching attributes in dataobjects
