### 2026-02-05 f8ee649
Supporting buffering metadata from front of file to avoid S3 traversals, and caching dataobjects instances, which were otherwise recreated several times.
