+++ pyfive/high_level.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 from collections import deque
 from collections.abc import Callable
 from collections.abc import Mapping, Sequence
+import io
 import os
 import posixpath
 import warnings
@@ -18,7 +19,7 @@ from pyfive.dataobjects import DataObjects, DatasetID
 from pyfive.misc_low_level import SuperBlock
 from pyfive.h5py import Datatype
 from pyfive.p5t import P5VlenStringType, P5ReferenceType, P5SequenceType
-
+from pyfive.utilities import MetadataBufferingWrapper
 
 class Group(Mapping):
     """
@@ -128,7 +129,7 @@ class Group(Mapping):
                 return None
 
         logging.info(f"[pyfive] Accessing object '{obj_name}' with link target {link_target} (lazy access: {noindex})")
-        dataobjs = DataObjects(self.file._fh, link_target)
+        dataobjs = self.file._get_dataobjects(link_target)
         if dataobjs.is_dataset:
             if additional_obj != ".":
                 raise KeyError("%s is a dataset, not a group" % (obj_name))
@@ -254,15 +255,32 @@ class File(Group):
         if hasattr(filename, "read"):
             if not hasattr(filename, "seek"):
                 raise ValueError("File like object must have a seek method")
-            self._fh = filename
+            fh = filename
             self.filename = getattr(filename, "name", "None")
         else:
-            self._fh = open(filename, "rb")
+            fh = open(filename, "rb")
             self._close = True
             self.filename = filename
+        
+        # Wrap S3 file handles with metadata buffering to reduce network calls
+        # Use adaptive buffer size: estimate from HDF5 superblock or use 1MB default
+        buffer_size = 1   # 1MB default
+        if isinstance(fh, MetadataBufferingWrapper):
+            # Already wrapped
+            self._fh = fh
+        elif type(fh).__name__ == "S3File" or hasattr(fh, "fs"):
+            # S3 file handle - wrap with buffering
+            # Try to estimate needed buffer from first few bytes if possible
+            logging.info("[pyfive] Detected S3 file, enabling metadata buffering (%d MB)", buffer_size)
+            self._fh = MetadataBufferingWrapper(fh, buffer_size=buffer_size)
+        else:
+            # Local file or other
+            self._fh = fh
+        
         self._superblock = SuperBlock(self._fh, 0)
+        self._dataobjects_cache = {}
         offset = self._superblock.offset_to_dataobjects
-        dataobjects = DataObjects(self._fh, offset)
+        dataobjects = self._get_dataobjects(offset)
 
         self.file = self
         self.mode = "r"
@@ -293,6 +311,15 @@ class File(Group):
     def __repr__(self) -> str:
         return '<HDF5 file "%s" (mode r)>' % (os.path.basename(self.filename))
 
+    def _get_dataobjects(self, obj_addr):
+        """Return cached DataObjects for an object header address."""
+        cached = self._dataobjects_cache.get(obj_addr)
+        if cached is not None:
+            return cached
+        dataobjects = DataObjects(self._fh, obj_addr)
+        self._dataobjects_cache[obj_addr] = dataobjects
+        return dataobjects
+
     def _get_object_by_address(self, obj_addr: BinaryIO) -> Self | Any | None:  # type: ignore[return]
         """Return the object pointed to by a given address."""
         if self._dataobjects.offset == obj_addr:
@@ -308,7 +335,7 @@ class File(Group):
                     # return instantiated object
                     return grp[name]
                 # descend only if it's a subgroup (need to instantiate minimally)
-                if DataObjects(self.file._fh, link_addr).is_group:
+                if self._get_dataobjects(link_addr).is_group:
                     queue.append((full_path, grp[name]))
 
     def close(self):
