### 2026-02-05 8805ad3
Documenting and linking the metadata buffer to File opening.
