### 2026-02-05 95868fd
restore environment.yml
