+++ environment.yml
@@ -14,7 +14,6 @@ dependencies:
   - flask-cors
   - h5netcdf
   - h5py  # link to same hdf5 lib with netcdf4
-  - lz4
   - moto
   - netcdf4 !=1.7.4  # github.com/Unidata/netcdf4-python/issues/1464
   - pre-commit  # with ruff and mypy
@@ -23,6 +22,7 @@ dependencies:
   - pytest-html
   - pytest-rerunfailures
   - pytest-xdist
+  - python-lzf
   - s3fs >=2025.9.0
   # documentation dependencies
   - autodocsumm
