### 2026-02-05 4adf545
add note in environment file
