### 2026-02-05 3544682
switch to new conda forge package
