+++ pyproject.toml
@@ -63,6 +63,7 @@ test = [
     "pytest-html",
     "pytest-rerunfailures",
     "pytest-xdist",
+    "python-neo-lzf",  # pypi.org/project/python-neo-lzf/
     "s3fs>=2025.9.0",  # v-mismatch with boto3 results in ancient s3fs
 ]
 
