### 2026-02-05 e3e66ac
remove unused lz4
