+++ pyproject.toml
@@ -54,7 +54,6 @@ test = [
     "flask-cors",
     "h5netcdf",
     "h5py",
-    "lz4",
     "moto",
     "netCDF4!=1.7.4",  # github.com/Unidata/netcdf4-python/issues/1464
     "pre-commit",  # with ruff and mypy
