### 2026-02-05 0052f18
Moving some logging to custom logger
