### 2026-02-06 #9688 @rich2355
Hi @bnlawrence, great job on the writing, I have no further comments there. Please complete the following so this can continue to move forward.

- Update Zenodo to the current version in Github, and update the DOIs.
- Correct and update the current reference list to include all DOIs. You can check your references with "@editorialbot check references." (There seem to have two refs without DOIs)
- Tell me what version number you would like to appear.

Once you complete these tasks, I can continue to move forward.

Thank you!

Best wishes,
RL
