### 2026-02-06 c6464bc
Further tidying up of logging
