+++ pyfive/h5d.py
@@ -413,7 +413,7 @@ class DatasetID:
         else:
             elapsed = f"{t1:.1f}s"
         logger.info(
-            "[pyfive] Chunk index built: btree range=%d; elapsed=%s",
+            "[pyfive] Chunk index built: btree range=%s; elapsed=%s",
             (self._btree_start,self._btree_end),
             elapsed,
         )
