### 2026-02-06 9e7a221
Avoid expensive logging actions if no logging happening
