+++ pyfive/dataobjects.py
@@ -10,6 +10,7 @@ import numpy as np
 from time import time
 import logging
 import inspect
+
 try:
     from importlib.metadata import version
 except ImportError:  # pragma: no cover
@@ -54,6 +55,7 @@ UNLIMITED_SIZE = UNDEFINED_ADDRESS
 
 logger = logging.getLogger(__name__)
 
+
 class DataObjects(object):
     """
     HDF5 DataObjects.
@@ -64,8 +66,8 @@ class DataObjects(object):
         # Log file handle info for diagnostics
         fh_id = id(fh)
         fh_type = type(fh).__name__
-        is_s3 = hasattr(fh, 'fs') or 'S3' in fh_type
-        
+        is_s3 = hasattr(fh, "fs") or "S3" in fh_type
+
         fh.seek(offset)
         version_hint = struct.unpack_from("<B", fh.read(1))[0]
         fh.seek(offset)
@@ -78,7 +80,10 @@ class DataObjects(object):
 
         logger.debug(
             "[pyfive] DataObjects init: fh_id=%s type=%s s3=%s offset=%d",
-            fh_id, fh_type, is_s3, offset
+            fh_id,
+            fh_type,
+            is_s3,
+            offset,
         )
 
         self.fh = fh
@@ -182,9 +187,9 @@ class DataObjects(object):
         """Return a dictionary of all attributes."""
         # Return cached attributes if available
         if self._cached_attributes is not None:
-            logger.debug('[pyfive] Attribute cache hit for offset %d', self.offset)
+            logger.debug("[pyfive] Attribute cache hit for offset %d", self.offset)
             return self._cached_attributes
-        
+
         t0 = time()
         attrs = {}
         attr_msgs = self.find_msg_type(ATTRIBUTE_MSG_TYPE)
@@ -197,30 +202,35 @@ class DataObjects(object):
         # Attributes may also be stored in objects reference in the
         # Attribute Info Message (0x0015, 21).
         # Assume we can have both types though I suspect this is not the case
-        attrs_log = f'({len(attrs)}'
+        attrs_log = f"({len(attrs)}"
         attr_info = self.find_msg_type(ATTRIBUTE_INFO_MSG_TYPE)
         if attr_info:
             more_attrs = self._get_attributes_from_attr_info(attrs, attr_info)
             attrs.update(more_attrs)
-            attrs_log += f'+{len(more_attrs)})'
+            attrs_log += f"+{len(more_attrs)})"
         else:
             attrs_log += ")"
         t1 = time() - t0
         if logger.isEnabledFor(logging.DEBUG):
-            pyfive_stack = [f for f in inspect.stack() if 'pyfive' in f.filename]
+            pyfive_stack = [f for f in inspect.stack() if "pyfive" in f.filename]
             if len(pyfive_stack) > 1:
                 logger.debug(
                     "[pyfive] stack: %s",
-                    " -> ".join(f"{f.function}" for f in pyfive_stack[1:])
+                    " -> ".join(f"{f.function}" for f in pyfive_stack[1:]),
                 )
         if offsets and logger.isEnabledFor(logging.INFO):
             fh_id = id(self.fh)
             fh_type = type(self.fh).__name__
             logger.info(
-                '[pyfive] Obtained %d%s attributes from offset %d (fh_id=%s type=%s) in %.4fs',
-                len(attrs), attrs_log, offsets[0], fh_id, fh_type, t1
+                "[pyfive] Obtained %d%s attributes from offset %d (fh_id=%s type=%s) in %.4fs",
+                len(attrs),
+                attrs_log,
+                offsets[0],
+                fh_id,
+                fh_type,
+                t1,
             )
-        
+
         # Cache the parsed attributes for subsequent calls
[…]
