+++ pyfive/dataobjects.py
@@ -11,11 +11,7 @@ from time import time
 import logging
 import inspect
 
-try:
-    from importlib.metadata import version
-except ImportError:  # pragma: no cover
-    from importlib_metadata import version
-
+from importlib.metadata import version
 from pyfive.datatype_msg import DatatypeMessage
 from pyfive.core import _padded_size, _structure_size
 from pyfive.core import _unpack_struct_from, _unpack_struct_from_file
