### 2026-02-06 183cd11
Fixing type issues and copilots criticisms.
