### 2026-02-06 #9688 @valeriupredoi
Hi @rich2355 great many thanks!

- about missing DOIs: both the references that come up as missing DOIs: "NetCDF-4: Software Implementing an Enhanced Data Model for the Geosciences" and "Chunking Data: Choosing Shapes" (blog post) don't actually have DOIs - well, the latter is a blog post anyway
- we've not posted the draft on Zenodo - do we have to? I'm not sure what you meant by "Update Zenodo to the current version in Github" my apologies

Cheers very much 🍻
