### 2026-02-06 #9688 @rich2355
Thanks @valeriupredoi!

- **Re missing DOIs**: If something has no DOIs, it is fine. It seems to that the two refs with missing DOIs actually have no DOIs, correct?
- **Re Zenodo**: This is a repository website and yes, you need to upload your project there.

I attached the guide for us here:

> At this point could you:
> - Make a tagged release of your software, and list the version tag of the archived version here.
> - Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> - Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> - Please list the DOI of the archived version here.
