### 2026-02-06 #9688 @valeriupredoi
@rich2355 cheers muchly!

> Re missing DOIs: If something has no DOIs, it is fine. It seems to that the two refs with missing DOIs actually have no DOIs, correct?

Indeed so, I double-checked that, they really don't have DOIs, unfortunately

> Re Zenodo: This is a repository website and yes, you need to upload your project there.

We will do that, next week!

Many thanks! 🍺
