### 2026-02-06 #9688 @rich2355
Thanks @valeriupredoi! No worries, the DOI questions are no longer questionable if they do not have DOIs.
