+++ Changelog.rst
@@ -3,14 +3,14 @@ Version 1.1.0
 
 **2026-02-10**
 
-### Highlights
+**Highlights**
 
 * Add support for pre-commit with mypy and Ruff and start adding type hints by `Valeriu Predoi <https://github.com/valeriupredoi>`_ in https://github.com/NCAS-CMS/pyfive/pull/173
 * Timing + Optimisations + Minor fixes by `Bryan Lawrence <https://github.com/bnlawrence>`_ in https://github.com/NCAS-CMS/pyfive/pull/191
 * Add ``__orthogonal_indexing__`` flag by `David Hassell <https://github.com/davidhassell>`_ in https://github.com/NCAS-CMS/pyfive/issue/171
 * Allow negative slices when indexing chunked data by `David Hassell <https://github.com/davidhassell>`_ in https://github.com/NCAS-CMS/pyfive/pull/170
 
-### Other Changes
+**Other Changes**
 
 * [JOSS REVIEW PR] Harmonize depenedncies from environment.yml and pyproject.toml by `Valeriu Predoi <https://github.com/valeriupredoi>`_ in https://github.com/NCAS-CMS/pyfive/pull/174
 * pep639 license by @valeriupredoi in https://github.com/NCAS-CMS/pyfive/pull/180
