### 2026-02-10 #9688 @valeriupredoi
hi @rich2355 we've now released pyfive=1.1.0 that holds all the review suggestions:

- GitHub: https://github.com/NCAS-CMS/pyfive/releases/tag/v1.1.0
- PyPI: https://pypi.org/project/pyfive/
- conda-forge: https://anaconda.org/channels/conda-forge/packages/pyfive/overview

I am now going to upload this new version to Zeonodo 🍻
