### 2026-02-10 #9688 @valeriupredoi
The Zenodo upload is done!

https://zenodo.org/records/18599472 with DOI https://doi.org/10.5281/zenodo.18599472

@rich2355 anything else we should be doing? 🍺
