### 2026-02-10 #9688 @rich2355
Thanks @valeriupredoi! Could you double check the zenodo link? That link seems to not work on my side...

In addition, could you set the DOI to be a hyperlink?
