### 2026-02-10 #9688 @valeriupredoi
> Thanks [@valeriupredoi](https://github.com/valeriupredoi)! Could you double check the zenodo link? That link seems to not work on my side...
>
> In addition, could you set the DOI to be a hyperlink?

sure! Here you go:

- Zenodo: https://zenodo.org/records/18599472
- URL to DOI: https://doi.org/10.5281/zenodo.18599472

The Zenodo link works well for me, from multiple devices, what error do you get? From a terminal:
```
$ firefox https://zenodo.org/records/18599472
```
-> all fine!
