### 2026-02-10 #9688 @valeriupredoi
I even popped a DOI badge on our GH 😁 https://github.com/NCAS-CMS/pyfive
