### 2026-02-10 #9688 @rich2355
Hi @valeriupredoi, there is only one item that I think may need more work:

> Check that the title, author list, version tag, and the license are correct in the archived release

So far, I think at least the author list and the license in JOSS paper is different from the one in the zenodo archive. Can you please check on that? Thanks!
