### 2026-02-11 #9688 @valeriupredoi
> Hi [@valeriupredoi](https://github.com/valeriupredoi), there is only one item that I think may need more work:
>
> > Check that the title, author list, version tag, and the license are correct in the archived release
>
> So far, I think at least the author list and the license in JOSS paper is different from the one in the zenodo archive. Can you please check on that? Thanks!

Hi @rich2355 - absolutely spot-on! Same Zenodo ref https://zenodo.org/records/18599472 but indeed I:

- updated the authors' list to be the same as the one in the paper (Wout De Nolf appears to have an ORCID number, but we didn't list in the paper, so I kept it like that on Zenodo too) - so 9 out 9 authors listed!
- updated License and added Copyright as per the LICENSE file
- cleaned up links in general description

I think it looks good now 😊 What you reckon? 🍻
