### 2026-02-11 f50a793
minor changes in paper.md
