+++ paper.md
@@ -68,20 +68,20 @@ With these extensions, coupled with thread safety, many of the limitations precl
 
 # Statement of need
 
-HDF5[^1][@FolEA11] is arguably the most important data format in physical science.
-It is of particular importance in the environmental sciences which rely on the netCDF4[^2][@Rew06] data format, which itself uses the HDF data format underneath. 
-From satellite missions, to climate models and radar systems, the default binary format has been HDF5 for decades. 
-While newer data formats are starting to get mindshare, there are petabytes, if not exabytes, of existing HDF5, and there remain many good use-cases for creating new data in the HDF5 format today. 
+HDF5[^1] [@FolEA11] is arguably the most important data format in physical science.
+It is of particular importance in the environmental sciences that rely on the netCDF4[^2] [@Rew06] data format, which itself uses the HDF data format underneath. 
+From satellite missions to climate models and radar systems, the default binary format has been HDF5 for decades. 
+While newer data formats are starting to get mindshare, there are petabytes, if not exabytes, of existing HDF5, and there remain many good use cases for creating new data in the HDF5 format today. 
 However, despite its historical importance, there are few libraries available for reading HDF5 file data that do not depend on the official HDF5 library maintained by the HDF Group. In particular, apart from `pyfive`, there are no Python HDF5 libraries that address the data access needs of environmental science. 
-While the HDF5 c-library is reliable and performant, and battle-tested over decades, there are some caveats to depending upon it: 
+While the HDF5 C library is reliable and performant, and battle-tested over decades, there are some caveats to depending upon it. 
 Firstly, it is not thread-safe. Secondly, the underlying code is large and complex, and should anything happen to the financial stability of the HDF5 Group, it is not obvious it could be maintained. 
-Finally, the code complexity also meant that it is not suitable for developing bespoke code for data recovery in the case of partially corrupt data. 
+Finally, the code complexity also means that it is not suitable for developing bespoke code for data recovery in the case of partially corrupt data. 
 From a long-term curation perspective these last two constraints present a major concern. 
 
 Reliance on a complex codebase controlled by a single private company presents significant challenges for long-term data access. Addressing these challenges requires well-documented data formats, the use of only those documented features, and the existence of publicly available code that can be sustainably maintained.
-The HDF Group have provided good documentation for the HDF5 format, but while there are communities of developers beyond those of the HDF Group, recent events suggest that given most of those developers and their existing funding are based in the USA, some spreading of risk would be desirable. 
-To that end, a pure Python code covering the core HDF5 features of interest to the target scientific community, which is relatively small and maintained by an international constituency provides some assurance that the community can maintain HDF5 access for the foreseeable future. 
-A pure Python code also makes it easier to develop scripts which can work around data and metadata corruption should they occur, and has the additional advantage of being able to be deployed in resource or operating-system constrained environments (such as on mobile).
+The HDF Group have provided good documentation for the HDF5 format, but while there are communities of developers beyond those of the HDF Group, recent events suggest that given most of those developers and their existing funding are based in the USA, and some spreading of risk would be desirable. 
+To that end, a pure Python code covering the core HDF5 features of interest to the target scientific community, which is relatively small and maintained by an international constituency, provides some assurance that the community can maintain HDF5 access for the foreseeable future. 
+A pure Python code also makes it easier to develop scripts that can work around data and metadata corruption should they occur, and has the additional advantage of being able to be deployed in resource or operating-system constrained environments (such as on mobile).
 
 
 [^1]: https://www.hdfgroup.org/solutions/hdf5/
[…]
