### 2026-02-11 #9688 @danielskatz
@bnlawrence - As track editor, I've proofread your submission, and suggest the changes in https://github.com/NCAS-CMS/pyfive/pull/195 - please merge this, or let me know what you disagree with.

Also, in the paper.md file, can you update affiliation 8 to "Independent Researcher" and add the appropriate country?
