### 2026-02-11 6006d68
Update paper.md
Co-authored-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
