### 2026-02-12 #9688 @bnlawrence
@danielskatz Thanks. I've approved all those except for one word, I think the "and" that I commented on doesn't make sense to me.

@valeriupredoi Can you fix the independent researcher bit?
