### 2026-02-12 6aa7b6a
change affiliation 8 to Independent Resercher, add country, and also add country to Astral Software Inc., both USA
