### 2026-02-12 112fd97
remove and from line 82
