### 2026-02-12 #9688 @valeriupredoi
> [@danielskatz](https://github.com/danielskatz) Thanks. I've approved all those except for one word, I think the "and" that I commented on doesn't make sense to me.
>
> [@valeriupredoi](https://github.com/valeriupredoi) Can you fix the independent researcher bit?

@danielskatz great many thanks for your suggestions! @bnlawrence and I have popped them all in https://github.com/NCAS-CMS/pyfive/pull/195 and merged the PR 🍻
