### 2026-02-12 #9688 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
