### 2026-02-12 85f8882
add country for affiliation 6
