+++ paper.md
@@ -45,7 +45,7 @@ affiliations:
    index: 4
  - name: Astral Software Inc., USA.
    index: 5
- - name: NIST Center for Neutron Research
+ - name: NIST Center for Neutron Research, USA.
    index: 6
  - name: Institute of Geosciences, Meteorology Section, University of Bonn, Germany.
    index: 7
