### 2026-02-12 #9688 @valeriupredoi
> [@bnlawrence](https://github.com/bnlawrence) & [@valeriupredoi](https://github.com/valeriupredoi) - Can you also add the country for affiliation 6? Sorry for missing this earlier.

bugger, completely missed that one too, just added it in [85f88826c4a650535803a3ffd5ed97ba13dcec01](https://github.com/NCAS-CMS/pyfive/pull/95/changes/85f88826c4a650535803a3ffd5ed97ba13dcec01)
