### 2026-02-12 #9688 @valeriupredoi
> Hi [@valeriupredoi](https://github.com/valeriupredoi) [@bnlawrence](https://github.com/bnlawrence), I will defer to the AEiC for the final step, but I would like to congrats on your work on this and I would like to say many thanks to the three reviewers [@Zeitsperre](https://github.com/Zeitsperre) [@ayenpure](https://github.com/ayenpure) [@dostuffthatmatters](https://github.com/dostuffthatmatters) in this work! Everyone is so positive and productive in the whole review process, and I am super happy I have a great start (yes, this is my first submission as an editor!)

very many thanks @rich2355 and also great many thanks to the reviewers! 🍻 Please see what @danielskatz just mentioned about being merged - that's something not for us to do, I assume? 😃
