### 2026-02-12 #9688 @danielskatz
@valeriupredoi - I'm sorry, I misread your PR as being for the missing country to be added to the affiliations in the paper branch. I now see that it's to merge the full paper branch into main.  I'll go ahead and finalize the paper processing based on the paper branch, then if you want to merge that into main, you can.
