### 2026-02-12 #9688 @danielskatz
Congratulations to @bnlawrence (Bryan Lawrence) & @valeriupredoi & all co-authors on your JOSS publication!!

And thanks to @Zeitsperre, @ayenpure, and @dostuffthatmatters for reviewing, and to @rich2355 for editing (his first time)!
JOSS completely depends on volunteers - we couldn't do this without you
